Fix authorization

+++ task_manager/users/views.py
@@ -5,6 +5,17 @@ from django.views import View
 from django.contrib import messages
 from django.contrib.auth.models import User
 from task_manager.users.forms import UserForm
+from django.contrib.auth.mixins import LoginRequiredMixin
+
+
+class CustomLoginRequiredMixin(LoginRequiredMixin):  
+    def dispatch(self, request, *args, **kwargs):  
+        if not request.user.is_authenticated:
+            messages.add_message(request, messages.WARNING, "Вы не авторизованы! Пожалуйста, выполните вход.", 
+                                 'alert alert-danger alert-dismissible fade show')
+            return redirect(reverse('login'))  
+        return super().dispatch(request, *args, **kwargs)  
+    
 
 class UserIndexView(View): # ДОБАВИТЬ ВЫВОД ОШИБОК
     def get(self, request, *args, **kwargs):
@@ -17,12 +28,6 @@ class UserIndexView(View): # ДОБАВИТЬ ВЫВОД ОШИБОК
                       'messages': mssgs,
             },
         )
-
-from django.contrib.auth.forms import UserCreationForm
-from django.contrib import messages
-from django.shortcuts import render, redirect
-from django.urls import reverse
-from django.views import View
 
 class UserFormCreateView(View): # ТУТ НУЖНО УБРАТЬ ЧТЕНИЕ СООБЩЕНИЙ
     def get(self, request, *args, **kwargs):
@@ -42,10 +47,16 @@ class UserFormCreateView(View): # ТУТ НУЖНО УБРАТЬ ЧТЕНИЕ С
         messages.add_message(request, messages.WARNING, "Проверьте заполняемые поля")
         return render(request, 'users/create.html', {'form': form, 'messages': mssgs})
     
-class UserFormUpdateView(View):
+class UserFormUpdateView(CustomLoginRequiredMixin, View):
     def get(self, request, *args, **kwargs):
         user_id = kwargs.get("id")
         user = User.objects.get(id=user_id)
+
+        if request.user != user:
+            messages.add_message(request, messages.WARNING, "У вас нет прав для изменения другого пользователя.", 
+                                 'alert alert-danger alert-dismissible fade show')
+            return redirect(reverse('user_index'))
+            
         form = UserForm(instance=user)
         mssgs = messages.get_messages(request)
         return render(
@@ -55,6 +66,12 @@ class UserFormUpdateView(View):
     def post(self, request, *args, **kwargs):
         user_id = kwargs.get("id")
         user = User.objects.get(id=user_id)
+
+        if request.user != user:
+            messages.add_message(request, messages.WARNING, "У вас нет прав для изменения другого пользователя.", 
+                                 'alert alert-danger alert-dismissible fade show')
+            return redirect(reverse('user_index'))
+        
         form = UserForm(request.POST, instance=user)
         if form.is_valid():
             form.save()
@@ -66,11 +83,16 @@ class UserFormUpdateView(View):
             request, "users/update.html", {"form": form, "user_id": user_id}
         )
     
-class UserFormDeleteView(View):
+class UserFormDeleteView(CustomLoginRequiredMixin, View):
     def get(self, request, *args, **kwargs):
         user_id = kwargs.get("id")
         user = User.objects.get(id=user_id)
-        mssgs = messages.get_messages(request)
+
+        if request.user != user:
+            messages.add_message(request, messages.WARNING, "У вас нет прав для изменения другого пользователя.", 
+                                 'alert alert-danger alert-dismissible fade show')
+            return redirect(reverse('user_index'))
+        
         return render(
             request, "users/delete.html", {"user": user, "user_id": user_id}
         )
@@ -78,6 +100,12 @@ class UserFormDeleteView(View):
     def post(self, request, *args, **kwargs):
         user_id = kwargs.get("id")
         user = User.objects.get(id=user_id)
+        
+        if request.user != user:
+            messages.add_message(request, messages.WARNING, "У вас нет прав для изменения другого пользователя.", 
+                                 'alert alert-danger alert-dismissible fade show')
+            return redirect(reverse('user_index'))
+        
         if user:
             user.delete()
         messages.add_message(request, messages.SUCCESS, "Пользователь успешно удалён",
